+++ clang-tools-extra/clang-tidy/hooks/HooksTidyModule.cpp
@@ -27,6 +27,7 @@
 #include "FieldDelBufLenCheck.h"
 #include "FloatArithPureCheck.h"
 #include "FloatComparePureCheck.h"
+#include "FloatManipPureCheck.h"
 #include "FloatOnePureCheck.h"
 #include "FloatPureCheck.h"
 #include "GuardCalledCheck.h"
@@ -83,6 +84,8 @@ public:
         "hooks-float-arith-pure");
     CheckFactories.registerCheck<FloatComparePureCheck>(
         "hooks-float-compare-pure");
+    CheckFactories.registerCheck<FloatManipPureCheck>(
+        "hooks-float-manip-pure");
     CheckFactories.registerCheck<FloatOnePureCheck>(
         "hooks-float-one-pure");
     CheckFactories.registerCheck<FloatPureCheck>(